Removed the integer MathQuestions
The decimal MathQuestions can also handle the integer ones, which caused redundancy.
- also added DivisionDecimal2Numbers
