+++ src/controller/LevelCreator.java
@@ -3,9 +3,9 @@ package controller;
 import model.Level;
 import model.monsters.RegularMonster;
 import model.monsters.RegularMonsterType;
-import model.questions.Addition2Integers;
-import model.questions.AdditionManyIntegers;
-import model.questions.Subtraction2Integers;
+import model.questions.Addition2Numbers;
+import model.questions.AdditionManyNumbers;
+import model.questions.Subtraction2Numbers;
 
 import java.util.LinkedList;
 
@@ -32,13 +32,13 @@ public class LevelCreator {
      * Method that creates the levels
      */
     public void createLevels() {
-        Level level1 = new Level(1, new Addition2Integers(1,9,1,9),
+        Level level1 = new Level(1, new Addition2Numbers(1,9,0,1,9,0),
                                  new RegularMonster(false, RegularMonsterType.Goblin,5,"You see a goblin standing right infront of you, it doesn't seem that strong.",
                                          "Prepare to die human!"), 3);
-        Level level2 = new Level(2, new AdditionManyIntegers(2,13,2),
+        Level level2 = new Level(2, new AdditionManyNumbers(2,13,0,3),
                                  new RegularMonster(false, RegularMonsterType.Skeleton, 7, "A skeleton stands in the way, defeat it!",
                                          "*bone cracks*"), 17);
-        Level level3 = new Level(3, new Subtraction2Integers(4, 16, 4, 16, false),
+        Level level3 = new Level(3, new Subtraction2Numbers(4, 16,0, 4, 16,0, false),
                 new RegularMonster(false, RegularMonsterType.Warden, 13, "You see a rather short warden upfront. Should you get any closer?",
                 "Halt! Who goes there?"), 20);
