+++ src/model/questions/Multiplication2Numbers.java
@@ -8,7 +8,7 @@ import java.math.RoundingMode;
  * to generate the numbers and answers. BigDecimal is used instead of double to aid with the precision and rounding.
  * @author Mattias Bengtsson
  */
-public class Multiplication2Decimals extends MathQuestions {
+public class Multiplication2Numbers extends MathQuestions {
     private BigDecimal[] answers;
     private double number1LowerBound;
     private double number1UpperBound;
@@ -20,14 +20,16 @@ public class Multiplication2Decimals extends MathQuestions {
     private BigDecimal number2;
 
     /**
-     * Constructor that initializes the instance variables and generates the answers.
+     * Constructor that initializes the instance variables for the bounds and the number of decimal places.
      * @param number1LowerBound the lowest value the first number can have.
      * @param number1UpperBound the highest value the first number can have.
+     * @param numOfDecimalsNumber1 the number of decimal places for the first number.
      * @param number2LowerBound the lowest value the second number can have.
      * @param number2UpperBound the highest value the second number can have.
+     * @param numOfDecimalsNumber2 the number of decimal places for the second number.
      */
-    public Multiplication2Decimals(double number1LowerBound, double number1UpperBound, int numOfDecimalsNumber1,
-                                   double number2LowerBound, double number2UpperBound, int numOfDecimalsNumber2) {
+    public Multiplication2Numbers(double number1LowerBound, double number1UpperBound, int numOfDecimalsNumber1,
+                                  double number2LowerBound, double number2UpperBound, int numOfDecimalsNumber2) {
         super();
         this.number1LowerBound = number1LowerBound;
         this.number1UpperBound = number1UpperBound;
@@ -64,8 +66,7 @@ public class Multiplication2Decimals extends MathQuestions {
     }
 
     /**
-     * Generates the correct answer and 3 fake answers in the answer array. The answers are all unique. The number of
-     * decimal places is automatically the bigger of the two numbers that are added.
+     * Generates the correct answer and 3 fake answers in the answer array. The answers are all unique.
      */
     private void generateAnswers() {
         answers = createBigDecimalAnswerArray();
@@ -86,7 +87,7 @@ public class Multiplication2Decimals extends MathQuestions {
      */
     private BigDecimal createFakeAnswer() {
         BigDecimal fakeAnswer;
-        int numOfDecimalsAnswer = numOfDecimalsNumber1 + numOfDecimalsNumber2;
+        int numOfDecimalsAnswer = answers[getCorrectAnswerIndex()].scale();
         while (true) {
             fakeAnswer = randomBigDecimal(number1LowerBound * number2LowerBound,
                     number1UpperBound * number2UpperBound, numOfDecimalsAnswer);
