+++ src/model/questions/TestMathQuestions.java
@@ -28,14 +28,12 @@ public class TestMathQuestions {
         boolean isAnswered;
 
         // choose which type of question
-//        mathQuestion = new Addition2Integers(1, 9, -9, 0);
-//        mathQuestion = new Subtraction2Integers(1, 18, 1, 9, false);
-//        mathQuestion = new Multiplication2Integers(1, 9, 1, 9);
-//        mathQuestion = new DivisionFraction2Integers(10, 100, 2, 10);
-//        mathQuestion = new AdditionManyIntegers(-9, 9, 5);
-//        mathQuestion = new Addition2Decimals(0.1, 0.15, 2, 0, 2, 1);
-//        mathQuestion = new Multiplication2Decimals(1, 9, 2, 1, 9, 1);
-        mathQuestion = new Subtraction2Decimals(1, 9, 0, 1, 9, 0, false);
+//        mathQuestion = new Addition2Numbers(0.1, 9.9, 0, 0, 7, 0);
+//        mathQuestion = new Multiplication2Numbers(1, 9, 2, 1, 9, 1);
+//        mathQuestion = new Subtraction2Numbers(1, 9, 0, 1, 9, 0, false);
+//        mathQuestion = new AdditionManyNumbers(1, 9, 2, 3);
+//        mathQuestion = new DivisionFraction2Numbers(10, 100, 2, 10);
+        mathQuestion = new DivisionDecimal2Numbers(10, 100, 0, 2, 10, 0, 2);
 
         while (true) {
             mathQuestion.generateNewQuestion();
